Refactored generate_support, restricted typeguard
After merging #16, the tests began failing due to a typeguard error
checking that a dictionary actually produces key, val pairs as
a returned item in `generate_support` in `utils.py`. I used the
opportunity to refactor that call to only ever return a dict by
making it optional to pass in the shuffled indexes (only used
internally in generate_episode, anyway), saving a second call to
`torch.unique`. Upon finding an issue in typeguard that appeared
at the same time as their upstream release and our error, I restricted
the typeguard version to be 4.0 or lower.

Refs #17