Updated to use tox
This little change became much bigger in thinking about whether
or not we should enforce runtime type-checking. The upshot is that
we can run pyright to enforce our types in the public-facing API
in the CI. However, pyright is a little challenging to run without
using a test environment like tox, so I brought in some of the
best practices from hydra-zen and adapted them for our code, things
like:
isort (to enforce module imports occur a certain way)
codespell (fix common spelling mistakes)

...and it was also simple to add a test that the docs can be built
as well as the "format" and "enforce-format" environments that
can ensure that the code passes our formatting requirements (and
can easily be fixed with `tox -e format` if not.